refactor: move PoxAddress to src/chainstate/stacks/address.rs, and have BurnStateDB::get_pox_payout_addrs() return Option<Vec<TupleData>, u128> instead of Option<Vec<PoxAddress>, u128>. This makes more sense since the chainstate module depends on PoxAddress far more so than Clarity. So, have the BurnStateDB return the tuple representation of a PoxAddress, and have the chainstate module convert that tuple into a PoxAddress once needed.
