+++ clarity/src/vm/database/mod.rs
@@ -17,8 +17,8 @@
 use std::collections::HashMap;
 
 pub use self::clarity_db::{
-    BurnStateDB, ClarityDatabase, HeadersDB, PoxAddress, StoreType, NULL_BURN_STATE_DB,
-    NULL_HEADER_DB, STORE_CONTRACT_SRC_INTERFACE,
+    BurnStateDB, ClarityDatabase, HeadersDB, StoreType, NULL_BURN_STATE_DB, NULL_HEADER_DB,
+    STORE_CONTRACT_SRC_INTERFACE,
 };
 pub use self::clarity_store::MemoryBackingStore;
 pub use self::clarity_store::{ClarityBackingStore, SpecialCaseHandler};
