+++ src/chainstate/burn/operations/leader_block_commit.rs
@@ -32,6 +32,7 @@ use crate::chainstate::burn::operations::{
 use crate::chainstate::burn::ConsensusHash;
 use crate::chainstate::burn::Opcodes;
 use crate::chainstate::burn::SortitionId;
+use crate::chainstate::stacks::address::PoxAddress;
 use crate::chainstate::stacks::index::storage::TrieFileStorage;
 use crate::chainstate::stacks::{StacksPrivateKey, StacksPublicKey};
 use crate::codec::{write_next, Error as codec_error, StacksMessageCodec};
@@ -44,8 +45,6 @@ use stacks_common::address::AddressHashMode;
 use stacks_common::util::hash::to_hex;
 use stacks_common::util::log;
 use stacks_common::util::vrf::{VRFPrivateKey, VRFPublicKey, VRF};
-
-use clarity::vm::PoxAddress;
 
 // return type from parse_data below
 struct ParsedData {
