+++ src/chainstate/burn/operations/mod.rs
@@ -38,6 +38,7 @@ use crate::types::chainstate::VRFSeed;
 
 use crate::chainstate::burn::ConsensusHash;
 use crate::chainstate::burn::Opcodes;
+use crate::chainstate::stacks::address::PoxAddress;
 use crate::util_lib::db::DBConn;
 use crate::util_lib::db::DBTx;
 use crate::util_lib::db::Error as db_error;
@@ -45,8 +46,6 @@ use stacks_common::util::hash::Hash160;
 use stacks_common::util::hash::Sha512Trunc256Sum;
 use stacks_common::util::secp256k1::MessageSignature;
 use stacks_common::util::vrf::VRFPublicKey;
-
-use clarity::vm::PoxAddress;
 
 use crate::types::chainstate::BurnchainHeaderHash;
