+++ src/chainstate/stacks/boot/mod.rs
@@ -23,6 +23,7 @@ use crate::burnchains::bitcoin::address::BitcoinAddress;
 use crate::burnchains::Burnchain;
 use crate::burnchains::{Address, PoxConstants};
 use crate::chainstate::burn::db::sortdb::SortitionDB;
+use crate::chainstate::stacks::address::PoxAddress;
 use crate::chainstate::stacks::db::StacksChainState;
 use crate::chainstate::stacks::index::marf::MarfConnection;
 use crate::chainstate::stacks::Error;
@@ -40,7 +41,6 @@ use clarity::vm::types::{
     PrincipalData, QualifiedContractIdentifier, SequenceData, StandardPrincipalData, TupleData,
     TypeSignature, Value,
 };
-use clarity::vm::PoxAddress;
 use stacks_common::address::AddressHashMode;
 use stacks_common::util::hash::Hash160;
 
@@ -450,7 +450,6 @@ pub mod test {
     use crate::net::test::*;
     use clarity::vm::contracts::Contract;
     use clarity::vm::types::*;
-    use clarity::vm::PoxAddress;
     use stacks_common::util::hash::to_hex;
     use stacks_common::util::*;
