+++ src/chainstate/stacks/db/transactions.rs
@@ -1378,6 +1378,7 @@ pub mod test {
     use rand::Rng;
 
     use crate::burnchains::Address;
+    use crate::chainstate::stacks::address::PoxAddress;
     use crate::chainstate::stacks::db::test::*;
     use crate::chainstate::stacks::index::storage::*;
     use crate::chainstate::stacks::index::*;
@@ -1390,7 +1391,6 @@ pub mod test {
     use clarity::vm::test_util::UnitTestBurnStateDB;
     use clarity::vm::test_util::TEST_BURN_STATE_DB;
     use clarity::vm::types::*;
-    use clarity::vm::PoxAddress;
     use stacks_common::types::chainstate::SortitionId;
     use stacks_common::util::hash::*;
 
@@ -8042,7 +8042,7 @@ pub mod test {
                 &self,
                 height: u32,
                 sortition_id: &SortitionId,
-            ) -> Option<(Vec<PoxAddress>, u128)> {
+            ) -> Option<(Vec<TupleData>, u128)> {
                 None
             }
         }
@@ -8231,7 +8231,7 @@ pub mod test {
                 &self,
                 height: u32,
                 sortition_id: &SortitionId,
-            ) -> Option<(Vec<PoxAddress>, u128)> {
+            ) -> Option<(Vec<TupleData>, u128)> {
                 None
             }
         }
